Minor fixes

+++ src/Maths/Silk.NET.Maths/Scalar.Constants.cs
@@ -172,7 +172,7 @@ namespace Silk.NET.Maths
                 E = (T) (object) FloatE;
                 Pi = (T) (object) FloatPi;
                 Tau = (T) (object) FloatTau;
-                MaxValueOver2 = (T) (object) (float.MaxValue / 2);
+                MaxValueOver2 = (T) (object) (float) (float.MaxValue / 2);
             }
             else if (typeof(T) == typeof(double))
             {
@@ -193,7 +193,7 @@ namespace Silk.NET.Maths
 #else
                 Tau = (T) (object) Math.Tau;
 #endif
-                MaxValueOver2 = (T) (object) (double.MaxValue / 2);
+                MaxValueOver2 = (T) (object) (double) (double.MaxValue / 2);
             }
             else if (typeof(T) == typeof(decimal))
             {
@@ -210,7 +210,7 @@ namespace Silk.NET.Maths
                 E = (T) (object) (decimal) Math.E;
                 Pi = (T) (object) (decimal) Math.PI;
                 Tau = Scalar.Multiply(Pi, Two);
-                MaxValueOver2 = (T) (object) (decimal.MaxValue / 2);
+                MaxValueOver2 = (T) (object) (decimal) (decimal.MaxValue / 2);
             }
             else if (typeof(T) == typeof(short))
             {
@@ -227,7 +227,7 @@ namespace Silk.NET.Maths
                 E = (T) (object) (short) FloatE;
                 Pi = (T) (object) (short) FloatPi;
                 Tau = (T) (object) (short) FloatTau;
-                MaxValueOver2 = (T) (object) (short.MaxValue / 2 + 1);
+                MaxValueOver2 = (T) (object) (short) (short.MaxValue / 2 + 1);
             }
             else if (typeof(T) == typeof(ushort))
             {
@@ -244,7 +244,7 @@ namespace Silk.NET.Maths
                 E = (T) (object) (ushort) FloatE;
                 Pi = (T) (object) (ushort) FloatPi;
                 Tau = (T) (object) (ushort) FloatTau;
-                MaxValueOver2 = (T) (object) (ushort.MaxValue / 2 + 1);
+                MaxValueOver2 = (T) (object) (ushort) (ushort.MaxValue / 2 + 1);
             }
             else if (typeof(T) == typeof(sbyte))
             {
@@ -261,7 +261,7 @@ namespace Silk.NET.Maths
                 E = (T) (object) (sbyte) FloatE;
                 Pi = (T) (object) (sbyte) FloatPi;
                 Tau = (T) (object) (sbyte) FloatTau;
-                MaxValueOver2 = (T) (object) (sbyte.MaxValue / 2 + 1);
+                MaxValueOver2 = (T) (object) (sbyte) (sbyte.MaxValue / 2 + 1);
             }
             else if (typeof(T) == typeof(byte))
             {
@@ -278,7 +278,7 @@ namespace Silk.NET.Maths
                 E = (T) (object) (byte) FloatE;
                 Pi = (T) (object) (byte) FloatPi;
                 Tau = (T) (object) (byte) FloatTau;
-                MaxValueOver2 = (T) (object) (byte.MaxValue / 2 + 1);
+                MaxValueOver2 = (T) (object) (byte) (byte.MaxValue / 2 + 1);
             }
             else if (typeof(T) == typeof(int))
             {
@@ -295,7 +295,7 @@ namespace Silk.NET.Maths
                 E = (T) (object) (int) FloatE;
                 Pi = (T) (object) (int) FloatPi;
                 Tau = (T) (object) (int) FloatTau;
-                MaxValueOver2 = (T) (object) (int.MaxValue / 2 + 1);
+                MaxValueOver2 = (T) (object) (int) (int.MaxValue / 2 + 1);
             }
             else if (typeof(T) == typeof(uint))
             {
@@ -312,7 +312,7 @@ namespace Silk.NET.Maths
                 E = (T) (object) (uint) FloatE;
                 Pi = (T) (object) (uint) FloatPi;
                 Tau = (T) (object) (uint) FloatTau;
-                MaxValueOver2 = (T) (object) (uint.MaxValue / 2 + 1);
+                MaxValueOver2 = (T) (object) (uint) (uint.MaxValue / 2 + 1);
             }
             else if (typeof(T) == typeof(long))
             {
@@ -329,7 +329,7 @@ namespace Silk.NET.Maths
                 E = (T) (object) (long) FloatE;
                 Pi = (T) (object) (long) FloatPi;
                 Tau = (T) (object) (long) FloatTau;
-                MaxValueOver2 = (T) (object) (long.MaxValue / 2 + 1);
+                MaxValueOver2 = (T) (object) (long) (long.MaxValue / 2 + 1);
             }
             else if (typeof(T) == typeof(ulong))
             {
@@ -346,7 +346,7 @@ namespace Silk.NET.Maths
                 E = (T) (object) (ulong) FloatE;
                 Pi = (T) (object) (ulong) FloatPi;
                 Tau = (T) (object) (ulong) FloatTau;
-                MaxValueOver2 = (T) (object) (ulong.MaxValue / 2 + 1);
+                MaxValueOver2 = (T) (object) (ulong) (ulong.MaxValue / 2 + 1);
             }
             else if (typeof(T) == typeof(BigInteger))
             {
